The administrative interface was removed from influxDB
In version 1.3, the web admin interface is no longer available in InfluxDB.
The interface does not run on port 8083 and InfluxDB ignores the [admin]
section in the configuration file if that section is present.

Chronograf replaces the web admin interface with improved tooling for
querying data, writing data, and database management. See Chronograf’s
transition guide for more information.

https://docs.influxdata.com/influxdb/v1.3/administration/config/#admin